edit and delete working properly
