+++ src/app/views/rolodex_view.js
@@ -1,7 +1,9 @@
 import $ from 'jquery';
 import _ from 'underscore';
 import Backbone from 'backbone';
+import Rolodex from 'app/collections/rolodex';
 import ContactView from 'app/views/contact_view';
+import ContactDetailView from 'app/views/contact_detail_view';
 
 const RolodexView = Backbone.View.extend({
   initialize: function(options){
@@ -35,6 +37,26 @@ const RolodexView = Backbone.View.extend({
     return this;
   },
 
+  makeDetail: function(model){
+    $('#contact-details').empty();
+    var cardDetail = new ContactDetailView({
+      model: model
+    });
+    cardDetail.render();
+    this.listenTo(cardDetail, "editMe", this.editCard);
+    $('#contact-details').append(cardDetail.$el);
+    $('#contact-details').show();
+  },
+
+  editCard: function(cardModel) {
+  console.log("Editing a card");
+
+  this.input.name.val(cardModel.get("name"));
+  this.input.email.val(cardModel.get("email"));
+  this.input.phone.val(cardModel.get("phone"));
+  this.model.remove(cardModel);
+},
+
   addContact: function(contact) {
     var card = new ContactView({
       // el: $('#contact-cards'), <---still unsure why when el was specified (and direct li objects were added to ul tag), .contact-card tags could not be clicked independently
@@ -42,6 +64,7 @@ const RolodexView = Backbone.View.extend({
       input: this.input,
       collection: this.model
     });
+    this.listenTo(card, "deetsplz", this.makeDetail);
     this.cardList.push(card); //add each contact view to cardlist matrix
   },
